send only one email per boot for reboot patterns
The "send command..." message happens a LOT and will fill the buffer of the serial.
Before rebooting the board, a lot of messages will be sent to the serial and thus
treated for reboot pattern matching (therefore, a message would be sent multiple times).

This patch fixes this by limiting the number of mails sent for reboot pattern to only one
per boot.

Signed-off-by: Free Electrons <tux@tux-laptop.free-electrons.com>